Add files via upload
A Library Management System (LMS) is a software-based solution that helps manage and automate the daily operations of a library. It is designed to simplify the management of library resources, such as books, journals, and other materials, and provide an efficient way to catalogue, search, and borrow items.
The proposed project aims to design and develop a comprehensive Library Management System that will improve the overall efficiency and effectiveness of library operations. The system will provide a user-friendly interface for librarians, administrators, and patrons to manage and access library resources, track circulation, and generate reports.